[feature]better modularization of component Card
